Fix: `-n` after `--task-name` properly sets args (#34)
Small bugfix PR. When running reproduced_results_analyzer.py, if the -n NUM_RUNS option was specified after the --task-name TASK_NAME option, then it would overwrite task_name. (The code was checking if opt in '--task-name', and since "-n" is a substring of "--task-name"...) This PR fixes this behavior.