Fix Docker disk space check failing in the spawner (#400)
We check whether there's enough disk space available for Docker images
by checking the disk usage of the Docker root dir. However, the Docker
root is not present in the spawner container (since it piggybacks off of
the host's Docker install), so that check fails with a
FileNotFoundError. This change fixes this by catching that error and
checking the regular disk space instead.