Hotfix: actually skip repos if json doesn't exist (#419)
The Classifier was not correctly handling the case where an input repo
does not have a corresponding JSON file (which can happen if e.g.
PairFinder finds no job pairs for that repo). It would try to process
that repo anyway, causing it to error when it tried opening a
nonexistent JSON file.