Temporary hotfix: pin requests below 2.31 (#437)
There is an incompatibility between the version of docker-py we use and
the latest versions of requests (2.32 and above). We should update
docker-py to a later version, but there are several breaking changes
between our version and the latest, so that will take some time to sort
out. For now, we can pin requests to the latest compatible version while
we update docker-py.