Allow overwriting artifact diffs (#444)
Previously, calling DatabaseAPI.set_diff() twice on the same image tag
would insert a second copy of the diff into the database, which is both
unintuitive and makes activities like correcting diffs with errors in
them difficult (since get_diff() returns the older version). This PR
changes the API to use upsert/PUT operations instead of insert/POST
operations, which allow for overwriting, and updates the database
settings to allow for this.