Fix bug in gradle caching (#456)
Previously, overzealous regex was altering `gradle/gradle-build-action`
to `gradle/gradle --offline-build-action`, causing artifacts that use
that action to fail to cache. This commit fixes that issue by reworking
how we detect gradle commands to add --offline to. It also attempts to
add the --offline flag to the `arguments` input of
`gradle/gradle-build-action`.